ISSUE #100 I've now defaulted the HttpMethod to GET and added a test to confirm
