+++ src/SevenDigital.Api.Wrapper.Unit.Tests/EndpointResolution/OAuth/UrlSignerTests.cs
@@ -1,78 +1,58 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using FakeItEasy;
 using NUnit.Framework;
-using SevenDigital.Api.Schema.OAuth;
 using SevenDigital.Api.Wrapper.EndpointResolution.OAuth;
 
 namespace SevenDigital.Api.Wrapper.Unit.Tests.EndpointResolution.OAuth
 {
 	[TestFixture]
-	public class OAuthSignatureGeneratorTests
+	public class UrlSignerTests
 	{
+		private ISignatureGenerator _signatureGenerator;
+		private UrlSigner _urlSigner;
 		private const string CONSUMER_KEY = "key";
 		private const string CONSUMER_SECRET = "secret";
+		private const string FAKE_URL = "test";
+		private const string FAKE_USER_TOKEN = "token";
+		private const string FAKE_USER_SECRET = "secret";
 
-		[Test]
-		public void SignUrlAsString_escapes_those_stupid_plus_signs_and_other_evils_in_signature()
+		[SetUp]
+		public void SetUp()
 		{
-			const string url = "http://www.example.com?parameter=hello&again=there";
-			var oAuthSignatureInfo = new OAuthSignatureInfo
-			{
-				FullUrlToSign = url,
-				ConsumerCredentials = GetOAuthCredentials(),
-				UserAccessToken = new OAuthAccessToken(),
-				HttpMethod = "GET"
-			};
-
-			for (int i = 0; i < 50; i++)
-			{
-				var signedUrl = new OAuthSignatureGenerator().Sign(oAuthSignatureInfo);
-				var index = signedUrl.IndexOf("oauth_signature");
-				var signature = signedUrl.Substring(index + "oauth_signature".Length);
-
-				Assert.That(!signature.Contains("+"), "signature contains a '+' character and isn't being encoded properly");
-			}
+			_signatureGenerator = A.Fake<ISignatureGenerator>();
+			_urlSigner = new UrlSigner(_signatureGenerator);
 		}
 
 		[Test]
-		public void SignUrl_adds_oauth_signature()
+		public void SignGetUrl_passes_expected_data_to_signature_generator()
 		{
-			var url = "http://www.example.com?parameter=hello&again=there";
-			var oAuthSignatureInfo = new OAuthSignatureInfo
-			{
-				FullUrlToSign = url,
-				ConsumerCredentials = GetOAuthCredentials(),
-				UserAccessToken = new OAuthAccessToken(),
-				HttpMethod = "GET"
-			};
-			var signedUrl = new Uri(new OAuthSignatureGenerator().Sign(oAuthSignatureInfo));
-			Assert.That(signedUrl.Query.Contains("oauth_signature"));
+			_urlSigner.SignGetUrl(FAKE_URL, FAKE_USER_TOKEN, FAKE_USER_SECRET, GetOAuthCredentials());
+			A.CallTo(() => _signatureGenerator.Sign(A<OAuthSignatureInfo>.That.Matches(x => x.HttpMethod == "GET"))).MustHaveHappened();
+			A.CallTo(() => _signatureGenerator.Sign(A<OAuthSignatureInfo>.That.Matches(x => x.FullUrlToSign == FAKE_URL))).MustHaveHappened();
+			A.CallTo(() => _signatureGenerator.Sign(A<OAuthSignatureInfo>.That.Matches(x => x.UserAccessToken.Token == FAKE_USER_TOKEN))).MustHaveHappened();
+			A.CallTo(() => _signatureGenerator.Sign(A<OAuthSignatureInfo>.That.Matches(x => x.UserAccessToken.Secret == FAKE_USER_SECRET))).MustHaveHappened();
 		}
 
 		[Test]
-		public void Make_sure_oath_token_is_encoded_when_POSTing()
+		public void SignPostRequest_passes_expected_data_to_signature_generator()
 		{
-			var url = "http://www.example.com/post";
+			_urlSigner.SignPostRequest(FAKE_URL, FAKE_USER_TOKEN, FAKE_USER_SECRET, GetOAuthCredentials(), new Dictionary<string, string> { { "one", "1" } });
+			A.CallTo(() => _signatureGenerator.SignWithPostData(A<OAuthSignatureInfo>.That.Matches(x => x.HttpMethod == "POST"))).MustHaveHappened();
+			A.CallTo(() => _signatureGenerator.SignWithPostData(A<OAuthSignatureInfo>.That.Matches(x => x.FullUrlToSign == FAKE_URL))).MustHaveHappened();
+			A.CallTo(() => _signatureGenerator.SignWithPostData(A<OAuthSignatureInfo>.That.Matches(x => x.UserAccessToken.Token == FAKE_USER_TOKEN))).MustHaveHappened();
+			A.CallTo(() => _signatureGenerator.SignWithPostData(A<OAuthSignatureInfo>.That.Matches(x => x.UserAccessToken.Secret == FAKE_USER_SECRET))).MustHaveHappened();
+		}
 
-			var oAuthSignatureInfo = new OAuthSignatureInfo
+		[Test]
+		public void SignPostRequest_passes_post_data_to_signature_generator()
+		{
+			var postParameters = new Dictionary<string, string>
 			{
-				FullUrlToSign = url,
-				ConsumerCredentials = GetOAuthCredentials(),
-				HttpMethod = "POST",
-				PostData = new Dictionary<string, string>
-				{
-					{"one", "1"}
-				},
-				UserAccessToken = new OAuthAccessToken
-				{
-					Token = "token==",
-					Secret = "secret=="
-				}
+				{"one", "1"}
 			};
-			var signedUrl = new OAuthSignatureGenerator().SignWithPostData(oAuthSignatureInfo);
 
-			Assert.That(signedUrl["oauth_token"], Is.EqualTo("token%3D%3D"));
+			_urlSigner.SignPostRequest(FAKE_URL, FAKE_USER_TOKEN, FAKE_USER_SECRET, GetOAuthCredentials(), postParameters);
+			A.CallTo(() => _signatureGenerator.SignWithPostData(A<OAuthSignatureInfo>.That.Matches(x => x.PostData == postParameters))).MustHaveHappened();
 		}
 
 		private IOAuthCredentials GetOAuthCredentials()
