+++ src/SevenDigital.Api.Wrapper.Unit.Tests/SevenDigital.Api.Wrapper.Unit.Tests.csproj
@@ -64,8 +64,8 @@
     <Compile Include="Deserialisation\User\Payment\DeleteCardTests.cs" />
     <Compile Include="EndpointResolution\EndpointResolverTests.cs" />
     <Compile Include="EndpointResolution\OAuth\NonceFactoryConcurrenyTests.cs" />
-    <Compile Include="EndpointResolution\OAuth\UrlSignerTest.cs" />
     <Compile Include="EndpointResolution\OAuth\UrlSignerTests.cs" />
+    <Compile Include="EndpointResolution\OAuth\OAuthSignatureGeneratorTests.cs" />
     <Compile Include="Exceptions\ApiWebExceptionTests.cs" />
     <Compile Include="Exceptions\ExceptionFactoryTests.cs" />
     <Compile Include="Exceptions\InputParameterExceptionTests.cs" />
